Split producers
