+++ src/Confluent.Kafka/IConsumer.cs
@@ -672,7 +672,7 @@ namespace Confluent.Kafka
         ///     The current consumer group metadata associated with this consumer,
         ///     or null if a GroupId has not been specified for the consumer.
         ///     This metadata object should be passed to the transactional producer's
-        ///     <see cref="IProducer{K,V}.SendOffsetsToTransaction(IEnumerable{TopicPartitionOffset},IConsumerGroupMetadata,TimeSpan)"/>
+        ///     <see cref="IProducerKV{K,V}.SendOffsetsToTransaction(IEnumerable{TopicPartitionOffset},IConsumerGroupMetadata,TimeSpan)"/>
         ///     method.
         /// </summary>
         IConsumerGroupMetadata ConsumerGroupMetadata { get; }
